refactor: move server-harness to command file
